superceed previous s2-constructors-formatters
