+++ R/RcppExports.R
@@ -253,18 +253,6 @@ cpp_s2_cell_common_ancestor_level_agg <- function(cellId) {
     .Call(`_s2_cpp_s2_cell_common_ancestor_level_agg`, cellId)
 }
 
-cpp_s2_geog_point <- function(x, y) {
-    .Call(`_s2_cpp_s2_geog_point`, x, y)
-}
-
-cpp_s2_make_line <- function(x, y, featureId) {
-    .Call(`_s2_cpp_s2_make_line`, x, y, featureId)
-}
-
-cpp_s2_make_polygon <- function(x, y, featureId, ringId, oriented, check) {
-    .Call(`_s2_cpp_s2_make_polygon`, x, y, featureId, ringId, oriented, check)
-}
-
 s2_geography_from_wkb <- function(wkb, oriented, check) {
     .Call(`_s2_s2_geography_from_wkb`, wkb, oriented, check)
 }
